move helpers and gulp task to a package
